% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_run.R
\name{coloc_run}
\alias{coloc_run}
\title{Run Bayesian enumeration colocalization using Coloc}
\usage{
coloc_run(
  df,
  trait_col = trait,
  variant_col = rsid,
  beta_col = beta,
  se_col = se,
  samplesize_col = samplesize,
  maf_col = maf,
  type_col = type,
  case_prop_col = case_prop,
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 1e-05,
  ...
)
}
\arguments{
\item{df}{Dataframe containing summary statistics at a single locus for two traits in a "long" format, with one row per variant per trait.}

\item{trait_col}{Column containing trait names}

\item{variant_col}{Column containing unique variant identifiers (Eg. rsids, chr:pos)}

\item{beta_col}{Column containing effect estimates}

\item{se_col}{Column containing standard errors}

\item{samplesize_col}{Column containing sample sizes}

\item{maf_col}{Column containing minor allele frequencies}

\item{type_col}{Column containing the type of each trait ("quant" for quantitative traits, "cc" for binary traits)}

\item{case_prop_col}{Column containing the proportion of cases for case control studies; this column is ignored for quantitative traits}

\item{p1}{Prior probability a SNP is associated with trait 1, default 1e-4}

\item{p2}{Prior probability a SNP is associated with trait 2, default 1e-4}

\item{p12}{Prior probability a SNP is associated with both traits, default 1e-5}

\item{...}{Arguments passed to \code{\link[coloc:coloc.abf]{coloc::coloc.abf()}}}
}
\value{
A list containing coloc results.
\itemize{
\item \code{summary} is a named vector containing the number of snps, and the posterior probabilities of the 5 colocalization hypotheses
\item \code{results} is an annotated version of the input data containing log approximate Bayes Factors and posterior probability of each SNP being causal if H4 is true.
}
}
\description{
This function is a wrapper around \code{\link[coloc:coloc.abf]{coloc::coloc.abf()}} that takes a dataframe as input, and performs colocalization under the single-causal-variant assumption. Coloc was described in Giambartolomei et al. (PLOS Genetics 2014; \url{https://doi.org/10.1371/journal.pgen.1004383}).
}
\examples{
\dontrun{
coloc_run(locus_df)
}
}
\seealso{
Other colocalization: 
\code{\link{hyprcoloc_df}()}
}
\concept{colocalization}
\concept{genomics}
