% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_qq_df.R
\name{gg_qq_df}
\alias{gg_qq_df}
\title{Create a QQ plot}
\usage{
gg_qq_df(sumstats_df, pval_col = p_value, ...)
}
\arguments{
\item{sumstats_df}{Dataframe containing GWAS summary statistics}

\item{pval_col}{Name of p-value column}

\item{...}{Arguments passed to \link[ggfastman:fast_qq]{ggfastman::fast_qq}}
}
\value{
A ggplot2 object
}
\description{
This function is a wrapper around \link[ggfastman:fast_qq]{ggfastman::fast_qq} which allows for the creation of a QQ plot from a dataframe containing GWAS summary statistics.
}
\examples{
\dontrun{
gg_qq_df(sumstats_df)
}
}
\seealso{
Other plotting: 
\code{\link{gg_manhattan_df}()}
}
\concept{genomics}
\concept{plotting}
