% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levinmisc-package.R
\docType{package}
\name{levinmisc-package}
\alias{levinmisc}
\alias{levinmisc-package}
\title{levinmisc: Miscellaneous Convenience Functions}
\description{
A set of miscellaneous convenience functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mglev1n/levinmisc}
  \item \url{https://mglev1n.github.io/levinmisc/}
  \item Report bugs at \url{https://github.com/mglev1n/levinmisc/issues}
}

}
\author{
\strong{Maintainer}: Michael Levin \email{mglev1n@gmail.com}

}
\keyword{internal}
