% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magmar.R
\name{magmar}
\alias{magmar}
\title{Run MAGMA gene-based analysis}
\usage{
magmar(
  sumstats_df,
  pval_col = p_value,
  snp_col = rsid,
  samplesize_col = samplesize,
  magma_bin,
  bfile,
  gene_file,
  out_file
)
}
\arguments{
\item{sumstats_df}{Dataframe containing GWAS summary statistics}

\item{pval_col}{Column containing p-values}

\item{snp_col}{Column containing rsids}

\item{samplesize_col}{Column containing sample size}

\item{magma_bin}{Path to MAGMA binary}

\item{bfile}{Path to reference data in plink \code{bfile} format (path should not include the extensions)}

\item{gene_file}{Path to file containing gene locations}

\item{out_file}{Output directory + file prefix}
}
\value{
Path to MAGMA results file
}
\description{
This function provides a wrapper to MAGMA, which performs gene-based testing from GWAS summary statistics. Details of MAGMA described in de Leeuw et al. (PLoS Cumputational Biology 2015; \url{https://doi.org/10.1371/journal.pcbi.1004219}). The MAGMA binary and reference files are available from the Complex Trait Genetics lab (\url{https://ctg.cncr.nl/software/magma}).
}
\seealso{
Other Gene-based testing: 
\code{\link{s_multixcan}()},
\code{\link{s_predixcan}()}
}
\concept{Gene-based testing}
\concept{genomics}
