% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal_config.R
\name{metal_config}
\alias{metal_config}
\title{Create a configuration file for METAL}
\usage{
metal_config(
  config_name,
  output_dir,
  study_files,
  SCHEME = "STDERR",
  AVERAGEFREQ = "ON",
  MINMAXFREQ = "OFF",
  TRACKPOSITIONS = "ON",
  MARKERLABEL = "MARKER",
  CHROMOSOMELABEL = "CHROM",
  POSITIONLABEL = "POS",
  EFFECT_ALLELE = "EFFECT_ALLELE",
  OTHER_ALLELE = "OTHER_ALLELE",
  EFFECTLABEL = "BETA",
  STDERR = "SE",
  FREQLABEL = "EAF",
  NCASE = "N_CASE",
  NCONTROL = "N_CONTROL",
  SAMPLESIZE = "N"
)
}
\arguments{
\item{config_name}{(string) Name of the configuration}

\item{output_dir}{(path) Path to output directory where GWAS meta-analysis data should be stored}

\item{study_files}{(list) List of paths to summary statistics that should be included in the GWAS}

\item{SCHEME}{(string) Either "SAMPLESIZE" or "STDERR" (default), corresponding to the METAL analysis scheme}

\item{AVERAGEFREQ}{(string) Either "ON" (default) or "OFF", allowing METAL to report the mean effect allele frequency across files}

\item{MINMAXFREQ}{(string) Either "ON" (default) or "OFF", allowing METAL to report the min/max effect allele frequency across files}

\item{TRACKPOSITIONS}{(string) Either "ON" (default) or "OFF", allowing METAL to report chromosome/position in the output}

\item{MARKERLABEL}{(string) Column containing unique markers to analyze (this column must be named the same across all input files)}

\item{CHROMOSOMELABEL}{(string) Column containing chromosomes (this column must be named the same across all input files)}

\item{POSITIONLABEL}{(string) Column containing genomic positions (this column must be named the same across all input files)}

\item{EFFECT_ALLELE}{(string) Column containing effect alleles (this column must be named the same across all input files)}

\item{OTHER_ALLELE}{(string) Column containing non-effect alleles (this column must be named the same across all input files)}

\item{EFFECTLABEL}{(string) Column containing effect sizes corresponding to the effect allele (this column must be named the same across all input files)}

\item{STDERR}{(string) Column containing standard errors fo the effect estimate (this column must be named the same across all input files)}

\item{FREQLABEL}{(string) Column containing effect allele frequencies (this column must be named the same across all input files)}

\item{NCASE}{(string) Column containing number of cases (this column must be named the same across all input files)}

\item{NCONTROL}{(string) Column containing number of controls (this column must be named the same across all input files)}

\item{SAMPLESIZE}{(string) Column containing total samplesize (this column must be named the same across all input files)}
}
\value{
Path to METAL configuration file
}
\description{
This function can be used to generate a configuration for METAL, a tool for performing meta-analysis of GWAS summary statistics \url{https://github.com/statgen/METAL}. The file created by this function can be used to run a meta-analysis. Details of the arguments to METAL are described in the METAL documentation: \url{https://genome.sph.umich.edu/wiki/METAL_Documentation}.
}
\seealso{
Run GWAS meta-analysis using METAL: \code{\link[=metal_run]{metal_run()}}

Other GWAS meta-analysis: 
\code{\link{metal_run}()},
\code{\link{mr_mega}()}
}
\concept{GWAS meta-analysis}
\concept{genomics}
