% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal_run.R
\name{metal_run}
\alias{metal_run}
\title{Use METAL to run a GWAS meta-analysis}
\usage{
metal_run(config_file, metal_path)
}
\arguments{
\item{config_file}{(path) Path to a METAL configuration file (this can be generated using \code{\link[=metal_config]{metal_config()}})}

\item{metal_path}{(path) Path to the METAL binary}
}
\value{
Path to gzip-formatted text file containing meta-analysis summary statistics
}
\description{
This function is a wrapper for METAL, a tool for performing meta-analysis of GWAS summary statistics \url{https://github.com/statgen/METAL}. Details of the arguments to METAL are described in the METAL documentation: \url{https://genome.sph.umich.edu/wiki/METAL_Documentation}.
}
\examples{
\dontrun{
metal_run(config_file = "config.txt", metal_path = "/path/to/metal_binary")
}
}
\seealso{
Create a METAL configuration file: \code{\link[=metal_config]{metal_config()}}

Other GWAS meta-analysis: 
\code{\link{metal_config}()},
\code{\link{mr_mega}()}
}
\concept{GWAS meta-analysis}
\concept{genomics}
