% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mega.R
\name{mr_mega}
\alias{mr_mega}
\title{Perform multi-ancestry GWAS meta-analysis using MR-MEGA}
\usage{
mr_mega(
  sumstats_files,
  mr_mega_bin,
  marker_col = MARKER,
  chr_col = CHROM,
  pos_col = POS,
  effect_allele_col = EFFECT_ALLELE,
  other_allele_col = OTHER_ALLELE,
  eaf_col = EAF,
  beta_col = BETA,
  se_col = SE,
  p_value_col = P,
  samplesize_col = N,
  n_pcs = 3
)
}
\arguments{
\item{sumstats_files}{List of files containing GWAS summary statistics. These files should all contain the same column header.}

\item{mr_mega_bin}{Path to MR-MEGA binary}

\item{marker_col}{Column containing unique markers for each variant}

\item{chr_col}{Column containing the chromosome}

\item{pos_col}{Column containing the position}

\item{effect_allele_col}{Column containing the effect allele}

\item{other_allele_col}{Column containing the non-effect allele}

\item{eaf_col}{Column containing effect allele frequencies}

\item{beta_col}{Column containing effect estimates}

\item{se_col}{Column containing standard errors}

\item{p_value_col}{Column containing p-value}

\item{samplesize_col}{Column containing sample size}

\item{n_pcs}{Number of genetic principal components to include in the meta-regression.}
}
\value{
A data.frame containing the GWAS summary statistics from the meta-regression
}
\description{
This function is a wrapper around MR-MEGA, which uses meta-regression to combine GWAS summart statistics while accounting for study/population-specific components of genetic variation. The method was described in Magi et al. (Human Molecular Genetics 2017; \url{https://doi.org/10.1093/hmg/ddx280}), and the package can be obtained from the Estonian Genome Centre (\url{https://genomics.ut.ee/en/tools}).
}
\examples{
\dontrun{
mr_mega(sumstats_files = list("/path/to/sumstats_1.txt.gz", "/path/to/sumstats_2.txt.gz"), mr_mega_bin = "/path/to/MR-MEGA")
}
}
\seealso{
Other GWAS meta-analysis: 
\code{\link{metal_config}()},
\code{\link{metal_run}()}
}
\concept{GWAS meta-analysis}
\concept{genomics}
