% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_multixcan.R
\name{s_multixcan}
\alias{s_multixcan}
\title{Integrate PrediXcan data across tissues}
\usage{
s_multixcan(
  df,
  models_folder,
  models_name_filter = ".*db",
  models_name_pattern,
  snp = SNP,
  effect_allele = effect_allele,
  other_allele = other_allele,
  beta = beta,
  eaf = eaf,
  chr = chr,
  pos = pos,
  se = se,
  pval = pval,
  samplesize = samplesize,
  regularization = 0.01,
  cutoff_condition_number = 30,
  cutoff_eigen_ratio = 0.001,
  cutoff_threshold = 0.4,
  cutoff_trace_ratio = 0.01,
  metaxcan_folder,
  metaxcan_filter,
  metaxcan_file_name_parse_pattern,
  snp_covariance,
  metaxcan,
  data
)
}
\arguments{
\item{df}{Dataframe containing GWAS summary statistics}

\item{models_folder}{Path to folder containing MetaXcan models (eg. \code{"MetaXcan/data/models/eqtl/mashr/"})}

\item{models_name_filter}{Filter used to identify model files (default = \code{".*\\\\.db"})}

\item{models_name_pattern}{Filter used to extract trait name from the model (default = \code{"mashr_(.*)\\\\.db"})}

\item{snp}{Column containing rsids}

\item{effect_allele}{Column containing effect alleles}

\item{other_allele}{Column containing non-effect alleles}

\item{beta}{Column containing effect estimates}

\item{eaf}{Column containing effect allele frequencies}

\item{chr}{Column containing chromosomes}

\item{pos}{Column containing positions}

\item{se}{Column containing standard errors}

\item{pval}{Column containing p-values}

\item{samplesize}{Column containing samplesize}

\item{regularization}{MetaXcan regularization (default = 0.01)}

\item{cutoff_condition_number}{Numer of eigen values to use when truncating SVD components (default = 30)}

\item{cutoff_eigen_ratio}{Ratio of eigenvalues to the max eigenvalue, as threshold to use when truncating SVD components. (default = 0.001)}

\item{cutoff_threshold}{Threshold of variance eigenvalues when truncating SVD (default = 0.4)}

\item{cutoff_trace_ratio}{Ratio of eigenvalues to trace, to use when truncating SVD (default = 0.01)}

\item{metaxcan_folder}{Path to folder containing S-PrediXcan result files}

\item{metaxcan_filter}{Regular expression to filter results files}

\item{metaxcan_file_name_parse_pattern}{Regular expression to get phenotype name and model name from MetaXcan result files. Assumes that a first group will be matched to phenotype name, and the second to model name.}

\item{snp_covariance}{Path to file containing S-MultiXcan covariance (eg. \code{"MetaXcan/data/models/gtex_v8_expression_mashr_snp_smultixcan_covariance.txt.gz"})}

\item{metaxcan}{Path to MetaXcan (eg. \code{"MetaXcan/software"})}

\item{data}{Path to MetaXcan data (eg. \code{"MetaXcan/data"})}
}
\value{
A dataframe containing S-MultiXcan results
}
\description{
This function is a wrapper around S-MultiXcan, a tool for identifying genes associate with a trait using GWAS summary statistics. This approach leverages the sharing of eQTLs acorss tissues to improve upon tissue-specific TWAS-methods like S-PrediXcan. This method uses data generated by \code{\link[=s_predixcan]{s_predixcan()}}. The S-MultiXcan method was described in Barbeira et al. (PLOS Genetics 2019; \url{https://doi.org/10.1371/journal.pgen.1007889}). The MetaXcan tools can be found on Github (\url{https://github.com/hakyimlab/MetaXcan}) and PredictDB (\url{https://predictdb.org/}).
}
\seealso{
Other TWAS: 
\code{\link{s_predixcan}()}
}
\concept{TWAS}
\concept{genomics}
