% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_predixcan.R
\name{s_predixcan}
\alias{s_predixcan}
\title{Run a TWAS using S-PrediXcan}
\usage{
s_predixcan(
  df,
  snp = SNP,
  effect_allele = effect_allele,
  other_allele = other_allele,
  beta = beta,
  eaf = eaf,
  chr = chr,
  pos = pos,
  se = se,
  pval = pval,
  samplesize = samplesize,
  data,
  metaxcan,
  output,
  model_db_path,
  model_covariance_path,
  trait_name
)
}
\arguments{
\item{df}{Dataframe containing GWAS summary statistics}

\item{snp}{Column containing rsid}

\item{effect_allele}{Column containing effect allele}

\item{other_allele}{Column containing non-effect allele}

\item{beta}{Column containing effect size}

\item{eaf}{Column containing effect allele frequency}

\item{chr}{Column containing chromosome}

\item{pos}{Column containing position}

\item{se}{Column containing standard error of the effect estimate}

\item{pval}{Column containing p-value}

\item{samplesize}{Column containing samplesize}

\item{data}{Path to MetaXcan data (eg. \code{"MetaXcan/data"})}

\item{metaxcan}{Path to MetaXcan (eg. \code{"MetaXcan/software"})}

\item{output}{Output directory to save S-PrediXcan results}

\item{model_db_path}{Path to PrediXcan model database}

\item{model_covariance_path}{Path to PrediXcan model covariance}

\item{trait_name}{Name of GWAS trait (used to name output files)}
}
\value{
A dataframe containing the S-PrediXcan results
}
\description{
This function is a wrapper around S-PrediXcan, a method of integrating GWAS summary statistics with gene-expression/splicing data to identify genes associated with a trait. The PrediXcan/MetaXcan method is described in Barbeira et al. (Nature Communications 2018; \url{https://doi.org/10.1038/s41467-018-03621-1}). The MetaXcan tools can be found on Github (\url{https://github.com/hakyimlab/MetaXcan}) and PredictDB (\url{https://predictdb.org/}). If S-PrediXcan is run across multiple tissues, the results can be integrated using \code{\link[=s_multixcan]{s_multixcan()}}.
}
\examples{
\dontrun{
s_predixcan(df, data = "MetaXcan/data", metaxcan = "MetaXcan/software", output = "/path/to/output", model_db_path = "MetaXcan/data/models/eqtl/mashr/mashr_Liver.db", model_covariance_path = "MetaXcan/data/models/eqtl/mashr/mashr_Liver.txt.gz", trait_name = "GWAS_trait")
}
}
\seealso{
Other TWAS: 
\code{\link{s_multixcan}()}

Other Gene-based testing: 
\code{\link{magmar}()},
\code{\link{s_multixcan}()}
}
\concept{Gene-based testing}
\concept{TWAS}
\concept{genomics}
