% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_rsids.R
\name{annotate_rsids}
\alias{annotate_rsids}
\title{Annotate dataframe with rsids}
\usage{
annotate_rsids(
  df,
  dbSNP = SNPlocs.Hsapiens.dbSNP144.GRCh37::SNPlocs.Hsapiens.dbSNP144.GRCh37,
  chrom_col = Chromosome,
  pos_col = Position
)
}
\arguments{
\item{df}{Dataframe containing variants to annotate with rsid}

\item{dbSNP}{Bioconductor object containing SNP locations and alleles to be used for annotation (default: \code{SNPlocs.Hsapiens.dbSNP144.GRCh37::SNPlocs.Hsapiens.dbSNP144.GRCh37})}

\item{chrom_col}{Chromosome column}

\item{pos_col}{Position column}
}
\description{
This function can be used rapidly add rsids to GWAS summary statistics by chromosome:position. This does not explicitly account for differences in variants at the position.
}
\examples{
\dontrun{
annotate_rsids()
}
}
\concept{genomics}
