% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc_df.R
\name{hyprcoloc_df}
\alias{hyprcoloc_df}
\title{Run multi-trait colocalization using HyPrColoc}
\usage{
hyprcoloc_df(
  df,
  trait_col = trait,
  snp_col = rsid,
  beta_col = beta,
  se_col = se,
  type_col = type,
  ...
)
}
\arguments{
\item{df}{Dataframe containing summary statistics at a single locus, in a "long" format, with one row per variant per trait.}

\item{trait_col}{Column containing trait names}

\item{snp_col}{Column containing variant names (eg. rsid, marker_name), which should be consistent across studies}

\item{beta_col}{Column containing effect estimates from GWAS}

\item{se_col}{Column containing standard errors of the effect estimates}

\item{type_col}{Column containing the "type" of trait - this column should contain \code{1} for binary traits, and \code{0} for all others}

\item{...}{Arguments passed to \code{\link[hyprcoloc:hyprcoloc]{hyprcoloc::hyprcoloc()}}}
}
\value{
A list containing a data.frame of HyPrColoc results: each row is a cluster of colocalized traits or is coded NA (if no colocalization is identified)
}
\description{
This function is a convenience wrapper around \code{\link[hyprcoloc:hyprcoloc]{hyprcoloc::hyprcoloc()}} that takes a dataframe as input, and performs mutli-trait colocalization. Details of the HyPrColoc method are described in Foley et al. (Nature Communications 2021; \url{https://doi.org/10.1038/s41467-020-20885-8}).
}
\examples{
\dontrun{
hyprcoloc_df(gwas_results_df)
}
}
\seealso{
Other {colocalization}: 
\code{\link{coloc_run}()}
}
\concept{genomics}
\concept{{colocalization}}
