% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldak_h2.R
\name{ldak_h2}
\alias{ldak_h2}
\title{Calculate heritability using LDAK}
\usage{
ldak_h2(
  sumstats_file,
  ldak_bin,
  ldak_tagfile,
  sample_prev = NULL,
  population_prev = NULL,
  hm3_file,
  ldak_cutoff = 0.01
)
}
\arguments{
\item{sumstats_file}{Path to "munged" GWAS summary statistics file in LDSC format. The current implementation of this function requires rsids to link to other LDAK files.}

\item{ldak_bin}{Path to LDAK binary}

\item{ldak_tagfile}{Path to LDAK tagfile}

\item{sample_prev}{Sample prevalence of cases (for case-control studies), \code{NULL} (default) for quantitative traits}

\item{population_prev}{Population prevalence of cases in the sample (for case-control studies), \code{NULL} (default) for quantitative traits}

\item{hm3_file}{Path to hm3 file containing}

\item{ldak_cutoff}{Minor allele frequency cutoff (default = 0.01)}
}
\value{
List of dataframes containing LDAK results
}
\description{
This function wraps LDAK, a command-line tool for estimating heritability. The tool and associated reference files can be download from the LDAK website (\url{https://ldak.org/}). The method is described in Zhang et al. (Nature Communications 2021; <https://doi.org/10.1038/s41467-021-24485-y).
}
\examples{
\dontrun{
ldak_h2(sumstats_file = "/path/to/munged_sumstats", ldak_bin = "/path/to/ldak", ldak_tagfile = "/pat/to/tagfile",  hm3_file = "/path/to/hm3")
}
}
\concept{genomics}
\concept{{heritability}}
