% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_extract_ld.R
\name{plink_extract_ld}
\alias{plink_extract_ld}
\title{Extract LD matrix using plink 1.9}
\usage{
plink_extract_ld(
  df,
  snp_col,
  effect_allele_col,
  metric = "r",
  bfile,
  threads = 1,
  memory = 16000,
  plink_bin
)
}
\arguments{
\item{df}{Dataframe containing variants to be included in the LD matrix; must contain SNP identifier and effect allele}

\item{snp_col}{Name of column containing SNP identifiers}

\item{effect_allele_col}{Name of column containing effect alleles (used when considering signed measures of LD like \code{r})}

\item{metric}{LD metric - either \code{r} (which will return a signed correlation matrix) or \code{r2} (which will return squared correlations)}

\item{bfile}{Path to plink \code{bfile} that will be used to extract LD}

\item{threads}{Number of threads (default = 1)}

\item{memory}{Memory limit (default = 16000 MB)}

\item{plink_bin}{Path to plink executable}
}
\value{
A square matrix containing the pairwise correlations between genetic variants
}
\description{
Extract LD matrix using plink 1.9
}
\examples{
\dontrun{
plink_extract_ld(df, snp_col, effect_allele_col, metric = "r", bfile = "/path/to/reference/panel/", plink_bin = "/path/to/plink1.9")
}
}
\concept{genomics}
\concept{{linkage}}
