% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_targets_proj.R
\name{populate_targets_proj}
\alias{populate_targets_proj}
\title{Populate Targets Project}
\usage{
populate_targets_proj(title, log_folder = "build_logs", overwrite = FALSE)
}
\arguments{
\item{title}{(character) base name for project files (eg. "{title}-Pipeline.qmd" and "{title}-Results.qmd")}

\item{log_folder}{(character) directory for LSF logs}

\item{overwrite}{(logical) overwrite existing template files}
}
\description{
Populate Targets Project
}
\examples{
\dontrun{
populate_targets_proj("test")
}
}
\concept{targets}
