% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_targets_proj.R
\name{populate_targets_proj}
\alias{populate_targets_proj}
\title{Create a minimal targets template in the current project}
\usage{
populate_targets_proj(title, log_folder = "build_logs", overwrite = FALSE)
}
\arguments{
\item{title}{(character) base name for project files (eg. "{title}-Pipeline.qmd" and "{title}-Results.qmd")}

\item{log_folder}{(character) directory for LSF logs}

\item{overwrite}{(logical) overwrite existing template files}
}
\description{
This function creates a minimal targets template in the current directory. This includes creating a \code{Pipelines.qmd} file containing boilerplate for running analyses, and a \code{Results.qmd} file which can be used to visualize the results. Parallelization of the pipeline is implemented using \code{\link[targets:tar_make]{targets::tar_make()}} and \code{crew.cluster}, using pre-filled using parameters specific to the LPC system at Penn.
}
\examples{
\dontrun{
populate_targets_proj("test")
}
}
\concept{targets}
