% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_datatable.R
\name{render_datatable}
\alias{render_datatable}
\title{Render DataTable to HTML}
\usage{
render_datatable(df)
}
\arguments{
\item{df}{Dataframe to render}
}
\value{
An HTML widget
}
\description{
This function is a wrapper around the \code{DT::datatable} function, containing useful defaults.
}
\examples{
render_datatable(datasets::mtcars)
}
\concept{miscellaneous}
