% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_carma.R
\name{run_carma}
\alias{run_carma}
\title{Perform Bayesian finemapping using CARMA}
\usage{
run_carma(
  df,
  snp_col,
  z_col,
  effect_allele_col,
  bfile,
  threads = 1,
  memory = 16000,
  plink_bin
)
}
\arguments{
\item{df}{Dataframe containing variants at a locus for finemapping}

\item{snp_col}{Name of column containing SNP identifiers}

\item{z_col}{Name of column containing signed Z-scores (relative to effect allele)}

\item{effect_allele_col}{Name of column containing effect alleles}

\item{bfile}{Path to plink \code{bfile} of reference panel that will be used to extract LD}

\item{threads}{Number of threads (default = 1)}

\item{memory}{Memory limit (default = 16000 MB)}

\item{plink_bin}{Path to plink executable}
}
\value{
A dataframe containing the input dataframe, and additional columns denoting which credible set (\code{CS}) each variant belongs to, as well as the posterior inclusion probability (\code{PIP}), and an \code{ld_error} column noting whether there were problems generating the LD matrix that limited fine-mapping.
}
\description{
This function is a wrapper around \code{CARMA::CARMA()} that takes a dataframe containing variants at a locus and performs Bayesian finemapping. The function requires a plink-formatted LD reference panel (\code{bfile}), which will be used to generate a signed LD matrix at the locus. CARMA was described in Yang et al. (Nature Genetics 2023; \url{https://doi.org/10.1038/s41588-023-01392-0})
}
\examples{
\dontrun{
run_carma(locus_df, snp_col = SNP, z_col = z, effect_allele_col = allele1)
}
}
\seealso{
Other {finemapping}: 
\code{\link{calc_credset}()}
}
\concept{genomics}
\concept{{finemapping}}
