% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_crew_lsf.R
\name{use_crew_lsf}
\alias{use_crew_lsf}
\title{Use Crew LSF}
\usage{
use_crew_lsf()
}
\value{
string
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function returns a template for using \code{crew.cluster} in a targets project, enabling the parallel execution of a targets workflow. By default, the template is pre-filled using parameters specific to the LPC system at Penn.
}
\examples{
\dontrun{
use_crew_lsf()
}
}
\concept{targets}
